# tyst-api-rest/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tyst_api_rest",
    "description": "TYST REST API",
    "license": {
      "name": "Apache-2.0 WITH FWM-Exception-1.0.0",
      "identifier": "Apache-2.0 WITH FWM-Exception-1.0.0"
    },
    "version": "0.0.0"
  },
  "paths": {
    "/digest/{algorithm}": {
      "post": {
        "tags": [
          "digest_resources"
        ],
        "summary": "Produce message digest (hash)",
        "description": "Produce message digest (hash) from binary input.\nSending \"Accept: */*\" returns the hex encoded version of the digest.",
        "operationId": "digest",
        "parameters": [
          {
            "name": "algorithm",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Message Digest (hash)",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              },
              "plain/text": {
                "schema": {
                  "type": "string"
                },
                "example": "0123456789abcdef0123456789abcdef (hex encoded)"
              }
            }
          },
          "400": {
            "description": "Failure"
          }
        }
      }
    },
...
  },
...
}